requirements.txt
openai
openai-agents
faiss-cpu
requests
numpy
